+++ model/groundwater.py
@@ -27,8 +27,7 @@ import subprocess
 import os
 import types
 
-import modelling_framework
-pcr, pcrfw = modelling_framework.load()
+from pcraster_provider import pcr
 
 import logging
 logger = logging.getLogger(__name__)
