+++ model/ncConverter.py
@@ -33,8 +33,7 @@ import glob
 import subprocess
 import netCDF4 as nc
 import numpy as np
-import modelling_framework
-pcr, pcrfw = modelling_framework.load()
+from pcraster_provider import pcr
 import virtualOS as vos
 
 # TODO: defined the dictionary (e.g. filecache = dict()) to avoid open and closing files
