+++ model/routing.py
@@ -30,8 +30,7 @@ import itertools
 
 from six.moves import map
 
-import modelling_framework
-pcr, pcrfw = modelling_framework.load()
+from pcraster_provider import pcr
 
 import logging
 logger = logging.getLogger(__name__)
