+++ model/waterBodies.py
@@ -25,8 +25,7 @@
 import os
 import types
 
-import modelling_framework
-pcr, pcrfw = modelling_framework.load()
+from pcraster_provider import pcr
 
 import logging
 logger = logging.getLogger(__name__)
